Extenshinoable architecture were added, fixes of access
