+++ headers/tools.h
@@ -1,0 +1,5 @@
+#include "dependencies.h"
+#ifndef APP_TOOLS_H
+#define APP_TOOLS_H
+tuple<string, string, string> argParser(int, char **);
+#endif //APP_TOOLS_H
